chap5-LCEL/pyproject.toml
[project]
name = "uv-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.11",
    "langchain_chroma==0.1.2",
    "langchain_community==0.2.11",
    "langchain_openai==0.1.21",
    "langchain_text_splitters==0.2.2",
    "langchain_core==0.2.29",
    "pypdf==4.3.1",
    "cryptography==43.0.0",
    "python-dotenv==1.0.1",
    "sudachipy==0.6.8",
    "sudachidict_core",
    "rank_bm25==0.2.2",
    "langgraph==0.2.16",
    "pdfplumber==0.11.4",
    "pymupdf==1.25.1",
    "numpy==1.26.4",
    "polars==1.18.0",
    "pandas==2.2.3",
    "open-clip-torch==2.29.0",
    "janome==0.5.0",
    "unstructured==0.16.12",
    "pymupdf4llm==0.0.17",
    "requests==2.32.3",
    "sentence-transformers==3.3.1",
    "markitdown==0.0.1a3",
    "py-zerox==0.0.7",
    "easyocr==1.7.2",
    "faiss-cpu==1.9.0.post1",
    "langsmith>=0.1.147",
    "tavily-python>=0.5.1",
]
